refactor: move color utils to common model folder
