+++ dev-client/src/model/color/soilColors.ts
