+++ dev-client/src/screens/SoilScreen/components/RenderValues.tsx
@@ -24,9 +24,9 @@ import {
 } from 'terraso-client-shared/soilId/soilIdSlice';
 
 import {Row, Text} from 'terraso-mobile-client/components/NativeBaseAdapters';
+import {munsellToString} from 'terraso-mobile-client/model/color/munsellConversions';
+import {isColorComplete} from 'terraso-mobile-client/model/color/soilColorValidation';
 import {ColorDisplay} from 'terraso-mobile-client/screens/SoilScreen/ColorScreen/components/ColorDisplay';
-import {munsellToString} from 'terraso-mobile-client/screens/SoilScreen/ColorScreen/utils/munsellConversions';
-import {isColorComplete} from 'terraso-mobile-client/screens/SoilScreen/ColorScreen/utils/soilColorValidation';
 
 export const renderDepth = (
   t: TFunction,
